+++ apps/dashboard/src/app/(dashboard)/(chain)/[chain_id]/[contractAddress]/_components/claim-conditions/snapshot-upload.tsx
@@ -1,43 +1,22 @@
 import { InlineCode } from "@/components/ui/inline-code";
-import { useThirdwebClient } from "@/constants/thirdweb.client";
+import { ToolTipLabel } from "@/components/ui/tooltip";
 import { cn } from "@/lib/utils";
 import {
   Box,
   Container,
   Flex,
-  IconButton,
   Link,
   ListItem,
-  Portal,
-  Select,
-  Table,
-  TableContainer,
-  Tbody,
-  Td,
-  Th,
-  Thead,
-  Tooltip,
-  Tr,
   UnorderedList,
 } from "@chakra-ui/react";
 import { Logo } from "components/logo";
-import {
-  ChevronFirstIcon,
-  ChevronLastIcon,
-  ChevronLeftIcon,
-  ChevronRightIcon,
-  CircleAlertIcon,
-  DownloadIcon,
-  UploadIcon,
-} from "lucide-react";
-import Papa from "papaparse";
-import { useCallback, useEffect, useRef, useState } from "react";
-import { type DropzoneOptions, useDropzone } from "react-dropzone";
-import { type Column, usePagination, useTable } from "react-table";
-import { isAddress } from "thirdweb";
-import { resolveAddress } from "thirdweb/extensions/ens";
+import { useCsvUpload } from "hooks/useCsvUpload";
+import { CircleAlertIcon, DownloadIcon, UploadIcon } from "lucide-react";
+import { useCallback, useRef } from "react";
+import type { Column } from "react-table";
+import { ZERO_ADDRESS } from "thirdweb";
 import { Button, Drawer, Heading, Text } from "tw-components";
-import { csvMimeTypes } from "utils/batch";
+import { CsvDataTable } from "../csv-data-table";
 
 interface SnapshotAddressInput {
   address: string;
@@ -50,150 +29,109 @@ interface SnapshotUploadProps {
   setSnapshot: (snapshot: SnapshotAddressInput[]) => void;
   isOpen: boolean;
   onClose: () => void;
-  value?:
-    | {
-        address: string;
-        maxClaimable: string;
-        price?: string | undefined;
-        currencyAddress?: string | undefined;
-      }[]
-    | null;
   dropType: "specific" | "any" | "overrides";
   isDisabled: boolean;
+  value?: SnapshotAddressInput[] | undefined;
 }
+
+const csvParser = (items: SnapshotAddressInput[]): SnapshotAddressInput[] => {
+  return items
+    .map(({ address, maxClaimable, price, currencyAddress }) => ({
+      address: (address || "").trim(),
+      maxClaimable: (maxClaimable || "0").trim().toLowerCase(),
+      price: (price || "").trim() || undefined,
+      currencyAddress: (currencyAddress || "").trim() || undefined,
+    }))
+    .filter(({ address }) => address !== "");
+};
 
 export const SnapshotUpload: React.FC<SnapshotUploadProps> = ({
   setSnapshot,
   isOpen,
   onClose,
-  value,
   dropType,
   isDisabled,
+  value,
 }) => {
-  const client = useThirdwebClient();
-  const [validSnapshot, setValidSnapshot] = useState<SnapshotAddressInput[]>(
-    value || [],
-  );
-  const [snapshotData, setSnapshotData] = useState<SnapshotAddressInput[]>([]);
-  const [noCsv, setNoCsv] = useState(false);
-  const [invalidFound, setInvalidFound] = useState(false);
-
-  const reset = useCallback(() => {
-    setValidSnapshot([]);
-    setNoCsv(false);
-  }, []);
+  const {
+    normalizeQuery,
+    getInputProps,
+    getRootProps,
+    isDragActive,
+    rawData,
+    noCsv,
+    reset,
+    removeInvalid,
+  } = useCsvUpload<SnapshotAddressInput>({ csvParser, defaultRawData: value });
 
   const _onClose = useCallback(() => {
     onClose();
   }, [onClose]);
 
-  const onDrop = useCallback<Required<DropzoneOptions>["onDrop"]>(
-    (acceptedFiles) => {
-      setNoCsv(false);
-
-      const csv = acceptedFiles.find(
-        (f) => csvMimeTypes.includes(f.type) || f.name?.endsWith(".csv"),
-      );
-      if (!csv) {
-        console.error(
-          "No valid CSV file found, make sure you have an address column.",
-        );
-        setNoCsv(true);
-        return;
-      }
-
-      Papa.parse(csv, {
-        header: true,
-        complete: (results) => {
-          const data: SnapshotAddressInput[] = (
-            results.data as SnapshotAddressInput[]
-          )
-            .map(({ address, maxClaimable, price, currencyAddress }) => ({
-              address: (address || "").trim(),
-              maxClaimable: (maxClaimable || "0").trim().toLowerCase(),
-              price: (price || "").trim() || undefined,
-              currencyAddress: (currencyAddress || "").trim() || undefined,
-            }))
-            .filter(({ address }) => address !== "");
-
-          if (!data[0]?.address) {
-            setNoCsv(true);
-            return;
-          }
-
-          setValidSnapshot(data);
-        },
-      });
-    },
-    [],
-  );
-
-  // FIXME: this can be a mutation or query instead!
-  // eslint-disable-next-line no-restricted-syntax
-  useEffect(() => {
-    if (validSnapshot.length === 0) {
-      return setSnapshotData([]);
-    }
-
-    const normalizeAddresses = async (snapshot: SnapshotAddressInput[]) => {
-      const normalized = await Promise.all(
-        snapshot.map(async ({ address, ...rest }) => {
-          let isValid = true;
-          let resolvedAddress = address;
-
-          try {
-            resolvedAddress = isAddress(address)
-              ? address
-              : await resolveAddress({ client, name: address });
-            isValid = !!resolvedAddress;
-          } catch {
-            isValid = false;
-          }
-
-          return {
-            address,
-            resolvedAddress,
-            isValid,
-            ...rest,
-          };
-        }),
-      );
-
-      const seen = new Set();
-      const filteredData = normalized.filter((el) => {
-        const duplicate = seen.has(el.resolvedAddress);
-        seen.add(el.resolvedAddress);
-        return !duplicate;
-      });
-
-      const valid = filteredData.filter(({ isValid }) => isValid);
-      const invalid = filteredData.filter(({ isValid }) => !isValid);
-
-      if (invalid?.length > 0) {
-        setInvalidFound(true);
-      }
-      const ordered = [...invalid, ...valid];
-      setSnapshotData(ordered);
-    };
-    normalizeAddresses(validSnapshot);
-  }, [validSnapshot, client]);
-
-  const removeInvalid = useCallback(() => {
-    const filteredData = snapshotData.filter(({ isValid }) => isValid);
-    setValidSnapshot(filteredData);
-    setInvalidFound(false);
-  }, [snapshotData]);
-
-  const { getRootProps, getInputProps, isDragActive } = useDropzone({
-    onDrop,
-  });
-
   const paginationPortalRef = useRef<HTMLDivElement>(null);
 
   const onSave = () => {
-    setSnapshot(snapshotData);
+    setSnapshot(normalizeQuery.data.result);
     onClose();
   };
+
+  const columns = [
+    {
+      Header: "Address",
+      accessor: ({ address, isValid }) => {
+        if (isValid) {
+          return address;
+        }
+        return (
+          <ToolTipLabel
+            label={
+              address === ZERO_ADDRESS
+                ? "Cannot send tokens to ZERO_ADDRESS"
+                : address.startsWith("0x")
+                  ? "Address is not valid"
+                  : "Address couldn't be resolved"
+            }
+          >
+            <div className="flex flex-row items-center gap-2">
+              <CircleAlertIcon className="size-4 text-red-500" />
+              <div className="cursor-default font-bold text-red-500">
+                {address}
+              </div>
+            </div>
+          </ToolTipLabel>
+        );
+      },
+    },
+    {
+      Header: "Max claimable",
+      accessor: ({ maxClaimable }) => {
+        return maxClaimable === "0" || !maxClaimable
+          ? "Default"
+          : maxClaimable === "unlimited"
+            ? "Unlimited"
+            : maxClaimable;
+      },
+    },
+    {
+      Header: "Price",
+      accessor: ({ price }) => {
+        return price === "0"
+          ? "Free"
+          : !price || price === "unlimited"
+            ? "Default"
+            : price;
+      },
+    },
+    {
+      Header: "Currency Address",
+      accessor: ({ currencyAddress }) => {
+        return currencyAddress ===
+          "0x0000000000000000000000000000000000000000" || !currencyAddress
+          ? "Default"
+          : currencyAddress;
+      },
+    },
+  ] as Column<SnapshotAddressInput>[];
 
   return (
     <Drawer
@@ -210,15 +148,19 @@ export const SnapshotUpload: React.FC<SnapshotUploadProps> = ({
               <div className="flex flex-row gap-2">
                 <Logo hideWordmark />
                 <Heading size="title.md">
-                  {validSnapshot.length ? "Edit" : "Upload"} Snapshot
+                  {rawData.length ? "Edit" : "Upload"} Snapshot
                 </Heading>
               </div>
             </Flex>
           </Container>
         </Flex>
 
-        {validSnapshot.length > 0 ? (
-          <SnapshotTable portalRef={paginationPortalRef} data={snapshotData} />
+        {rawData.length > 0 ? (
+          <CsvDataTable<SnapshotAddressInput>
+            portalRef={paginationPortalRef}
+            data={normalizeQuery.data.result}
+            columns={columns}
+          />
         ) : (
           <Flex flexGrow={1} align="center" overflow="auto">
             <Container maxW="container.page">
@@ -374,7 +316,7 @@ export const SnapshotUpload: React.FC<SnapshotUploadProps> = ({
               >
                 <Button
                   borderRadius="md"
-                  disabled={isDisabled || validSnapshot.length === 0}
+                  disabled={isDisabled || rawData.length === 0}
                   onClick={() => {
                     reset();
                   }}
@@ -382,11 +324,11 @@ export const SnapshotUpload: React.FC<SnapshotUploadProps> = ({
                 >
                   Reset
                 </Button>
-                {invalidFound ? (
+                {normalizeQuery.data.invalidFound ? (
                   <Button
                     borderRadius="md"
                     colorScheme="primary"
-                    disabled={isDisabled || validSnapshot.length === 0}
+                    disabled={isDisabled || rawData.length === 0}
                     onClick={() => {
                       removeInvalid();
                     }}
@@ -400,7 +342,7 @@ export const SnapshotUpload: React.FC<SnapshotUploadProps> = ({
                     colorScheme="primary"
                     onClick={onSave}
                     w={{ base: "100%", md: "auto" }}
-                    isDisabled={isDisabled || validSnapshot.length === 0}
+                    isDisabled={isDisabled || rawData.length === 0}
                   >
                     Next
                   </Button>
@@ -411,197 +353,5 @@ export const SnapshotUpload: React.FC<SnapshotUploadProps> = ({
         </Flex>
       </Flex>
     </Drawer>
-  );
-};
-
-interface SnapshotTableProps {
-  data: SnapshotAddressInput[];
-  portalRef: React.RefObject<HTMLDivElement | null>;
-}
-
-const SnapshotTableColumns = [
-  {
-    Header: "Address",
-    accessor: ({ address, isValid }) => {
-      if (isValid) {
-        return address;
-      }
-      return (
-        <div className="flex flex-row items-center gap-2">
-          <Tooltip
-            label={
-              address.startsWith("0x")
-                ? "Address is not valid"
-                : "Address couldn't be resolved"
-            }
-          >
-            <div className="flex flex-row items-center gap-2">
-              <CircleAlertIcon className="size-4 text-red-500" />
-              <Text fontWeight="bold" color="red.500" cursor="default">
-                {address}
-              </Text>
-            </div>
-          </Tooltip>
-        </div>
-      );
-    },
-  },
-  {
-    Header: "Max claimable",
-    accessor: ({ maxClaimable }) => {
-      return maxClaimable === "0" || !maxClaimable
-        ? "Default"
-        : maxClaimable === "unlimited"
-          ? "Unlimited"
-          : maxClaimable;
-    },
-  },
-  {
-    Header: "Price",
-    accessor: ({ price }) => {
-      return price === "0"
-        ? "Free"
-        : !price || price === "unlimited"
-          ? "Default"
-          : price;
-    },
-  },
-  {
-    Header: "Currency Address",
-    accessor: ({ currencyAddress }) => {
-      return currencyAddress === "0x0000000000000000000000000000000000000000" ||
-        !currencyAddress
-        ? "Default"
-        : currencyAddress;
-    },
-  },
-] as Column<SnapshotAddressInput>[];
-
-const SnapshotTable: React.FC<SnapshotTableProps> = ({ data, portalRef }) => {
-  const {
-    getTableProps,
-    getTableBodyProps,
-    headerGroups,
-    prepareRow,
-    // Instead of using 'rows', we'll use page,
-    page,
-    // which has only the rows for the active page
-
-    // The rest of these things are super handy, too ;)
-    canPreviousPage,
-    canNextPage,
-    pageOptions,
-    pageCount,
-    gotoPage,
-    nextPage,
-    previousPage,
-    setPageSize,
-    state: { pageIndex, pageSize },
-  } = useTable(
-    {
-      columns: SnapshotTableColumns,
-      data,
-      initialState: {
-        pageSize: 50,
-        pageIndex: 0,
-      },
-    },
-    // old package, this will be removed
-    // eslint-disable-next-line react-compiler/react-compiler
-    usePagination,
-  );
-
-  // Render the UI for your table
-  return (
-    <Flex flexGrow={1}>
-      <TableContainer>
-        <Table {...getTableProps()}>
-          <Thead>
-            {headerGroups.map((headerGroup, headerGroupIndex) => (
-              // biome-ignore lint/suspicious/noArrayIndexKey: FIXME
-              <Tr {...headerGroup.getHeaderGroupProps()} key={headerGroupIndex}>
-                {headerGroup.headers.map((column, columnIndex) => (
-                  // biome-ignore lint/suspicious/noArrayIndexKey: FIXME
-                  <Th {...column.getHeaderProps()} key={columnIndex}>
-                    <Text as="label" size="label.sm" color="faded">
-                      {column.render("Header")}
-                    </Text>
-                  </Th>
-                ))}
-              </Tr>
-            ))}
-          </Thead>
-          <Tbody {...getTableBodyProps()}>
-            {page.map((row, rowIndex) => {
-              prepareRow(row);
-              return (
-                // biome-ignore lint/suspicious/noArrayIndexKey: FIXME
-                <Tr {...row.getRowProps()} key={rowIndex}>
-                  {row.cells.map((cell, cellIndex) => {
-                    return (
-                      <Td
-                        {...cell.getCellProps()}
-                        borderColor="borderColor"
-                        // biome-ignore lint/suspicious/noArrayIndexKey: FIXME
-                        key={cellIndex}
-                      >
-                        {cell.render("Cell")}
-                      </Td>
-                    );
-                  })}
-                </Tr>
-              );
-            })}
-          </Tbody>
-        </Table>
-      </TableContainer>
-      <Portal containerRef={portalRef}>
-        <div className="flex w-full items-center justify-center">
-          <div className="flex flex-row gap-1">
-            <IconButton
-              isDisabled={!canPreviousPage}
-              aria-label="first page"
-              icon={<ChevronFirstIcon className="size-4" />}
-              onClick={() => gotoPage(0)}
-            />
-            <IconButton
-              isDisabled={!canPreviousPage}
-              aria-label="previous page"
-              icon={<ChevronLeftIcon className="size-4" />}
-              onClick={() => previousPage()}
-            />
-            <p className="my-auto whitespace-nowrap">
-              Page <strong>{pageIndex + 1}</strong> of{" "}
-              <strong>{pageOptions.length}</strong>
-            </p>
-            <IconButton
-              isDisabled={!canNextPage}
-              aria-label="next page"
-              icon={<ChevronRightIcon className="size-4" />}
-              onClick={() => nextPage()}
-            />
-            <IconButton
-              isDisabled={!canNextPage}
-              aria-label="last page"
-              icon={<ChevronLastIcon className="size-4" />}
-              onClick={() => gotoPage(pageCount - 1)}
-            />
-
-            <Select
-              onChange={(e) => {
-                setPageSize(Number.parseInt(e.target.value as string, 10));
-              }}
-              value={pageSize}
-            >
-              <option value="25">25</option>
-              <option value="50">50</option>
-              <option value="100">100</option>
-              <option value="250">250</option>
-              <option value="500">500</option>
-            </Select>
-          </div>
-        </div>
-      </Portal>
-    </Flex>
   );
 };
