+++ apps/dashboard/src/app/(dashboard)/(chain)/[chain_id]/[contractAddress]/tokens/components/airdrop-upload.tsx
@@ -1,37 +1,14 @@
 import { UnorderedList } from "@/components/ui/List/List";
 import { ToolTipLabel } from "@/components/ui/tooltip";
-import { useThirdwebClient } from "@/constants/thirdweb.client";
 import { cn } from "@/lib/utils";
-import {
-  IconButton,
-  Link,
-  Portal,
-  Select,
-  Table,
-  Tbody,
-  Td,
-  Th,
-  Thead,
-  Tr,
-} from "@chakra-ui/react";
-import { useQueries } from "@tanstack/react-query";
-import {
-  ChevronFirstIcon,
-  ChevronLastIcon,
-  ChevronLeftIcon,
-  ChevronRightIcon,
-  CircleAlertIcon,
-  UploadIcon,
-} from "lucide-react";
-import Papa from "papaparse";
-import { useCallback, useMemo, useRef, useState } from "react";
-import { type DropzoneOptions, useDropzone } from "react-dropzone";
-import { type Column, usePagination, useTable } from "react-table";
-import { type ThirdwebClient, ZERO_ADDRESS } from "thirdweb";
-import { resolveAddress } from "thirdweb/extensions/ens";
-import { isAddress } from "thirdweb/utils";
+import { Link } from "@chakra-ui/react";
+import { useCsvUpload } from "hooks/useCsvUpload";
+import { CircleAlertIcon, UploadIcon } from "lucide-react";
+import { useMemo, useRef } from "react";
+import type { Column } from "react-table";
+import { ZERO_ADDRESS } from "thirdweb";
 import { Button, Heading, Text } from "tw-components";
-import { csvMimeTypes } from "utils/batch";
+import { CsvDataTable } from "../../_components/csv-data-table";
 
 export interface AirdropAddressInput {
   address: string;
@@ -42,145 +19,89 @@ interface AirdropUploadProps {
   setAirdrop: (airdrop: AirdropAddressInput[]) => void;
   onClose: () => void;
 }
-async function checkIsAddress(
-  item: AirdropAddressInput,
-  thirdwebClient: ThirdwebClient,
-) {
-  const { address, ...rest } = item;
-  // Sending tokens to the zero address will result in an error
-  // if user wants to burn, they should use the Burn tab
-  if (address === ZERO_ADDRESS) {
-    return {
-      address,
-      resolvedAddress: address,
-      isValid: false,
-      ...rest,
-    };
-  }
-  let isValid = true;
-  let resolvedAddress = address;
-  try {
-    resolvedAddress = isAddress(address)
-      ? address
-      : await resolveAddress({ name: address, client: thirdwebClient });
-    isValid = !!resolvedAddress;
-  } catch {
-    isValid = false;
-  }
-  return {
-    address,
-    resolvedAddress,
-    isValid,
-    ...rest,
-  };
-}
-function processAirdropData(
-  data: ((AirdropAddressInput & { resolvedAddress: string }) | undefined)[],
-) {
-  const seen = new Set();
-  const filteredData = data
-    .filter((o) => o !== undefined)
-    .filter((el) => {
-      const duplicate = seen.has(el.resolvedAddress);
-      seen.add(el.resolvedAddress);
-      return !duplicate;
-    });
-  const valid = filteredData.filter(({ isValid }) => isValid);
-  const invalid = filteredData.filter(({ isValid }) => !isValid);
-  // Make sure the invalid records are at the top so that users can see
-  const ordered = [...invalid, ...valid];
-  return ordered;
-}
+
+const csvParser = (items: AirdropAddressInput[]): AirdropAddressInput[] => {
+  return items
+    .map(({ address, quantity }) => ({
+      address: (address || "").trim(),
+      quantity: (quantity || "1").trim(),
+    }))
+    .filter(({ address }) => address !== "");
+};
 
 export const AirdropUpload: React.FC<AirdropUploadProps> = ({
   setAirdrop,
   onClose,
 }) => {
-  const thirdwebClient = useThirdwebClient();
-  const [validAirdrop, setValidAirdrop] = useState<AirdropAddressInput[]>([]);
-  const [noCsv, setNoCsv] = useState(false);
-  const reset = useCallback(() => {
-    setValidAirdrop([]);
-    setNoCsv(false);
-  }, []);
-  const onDrop = useCallback<Required<DropzoneOptions>["onDrop"]>(
-    (acceptedFiles) => {
-      setNoCsv(false);
-      const csv = acceptedFiles.find(
-        (f) => csvMimeTypes.includes(f.type) || f.name?.endsWith(".csv"),
-      );
-      if (!csv) {
-        console.error(
-          "No valid CSV file found, make sure you have an address column.",
-        );
-        setNoCsv(true);
-        return;
-      }
-      Papa.parse(csv, {
-        header: true,
-        complete: (results) => {
-          const data: AirdropAddressInput[] = (
-            results.data as AirdropAddressInput[]
-          )
-            .map(({ address, quantity }) => ({
-              address: (address || "").trim(),
-              quantity: (quantity || "1").trim(),
-            }))
-            .filter(({ address }) => address !== "");
-          if (!data[0]?.address) {
-            setNoCsv(true);
-            return;
-          }
-          setValidAirdrop(data);
-        },
-      });
-    },
-    [],
-  );
-  const normalizeQuery = useQueries({
-    queries: validAirdrop.map((o) => ({
-      queryKey: ["snapshot-check-isAddress", o.address],
-      queryFn: () => checkIsAddress(o, thirdwebClient),
-    })),
-    combine: (results) => {
-      return {
-        data: {
-          result: processAirdropData(results.map((result) => result.data)),
-          invalidFound: !!results.find((o) => !o.data?.isValid),
-        },
-        pending: results.some((result) => result.isPending),
-      };
-    },
-  });
-  const { getRootProps, getInputProps, isDragActive } = useDropzone({
-    onDrop,
-  });
+  const {
+    normalizeQuery,
+    getInputProps,
+    getRootProps,
+    isDragActive,
+    rawData,
+    noCsv,
+    reset,
+    removeInvalid,
+  } = useCsvUpload<AirdropAddressInput>({ csvParser });
   const paginationPortalRef = useRef<HTMLDivElement>(null);
   const onSave = () => {
     setAirdrop(normalizeQuery.data.result);
     onClose();
   };
-  const removeInvalid = useCallback(() => {
-    const filteredData = normalizeQuery.data?.result.filter(
-      ({ isValid }) => isValid,
-    );
-    setValidAirdrop(filteredData);
-  }, [normalizeQuery.data.result]);
+
+  const columns = useMemo(() => {
+    return [
+      {
+        Header: "Address",
+        accessor: ({ address, isValid }) => {
+          if (isValid) {
+            return address;
+          }
+          return (
+            <ToolTipLabel
+              label={
+                address === ZERO_ADDRESS
+                  ? "Cannot send tokens to ZERO_ADDRESS"
+                  : address.startsWith("0x")
+                    ? "Address is not valid"
+                    : "Address couldn't be resolved"
+              }
+            >
+              <div className="flex flex-row items-center gap-2">
+                <CircleAlertIcon className="size-4 text-red-500" />
+                <div className="cursor-default font-bold text-red-500">
+                  {address}
+                </div>
+              </div>
+            </ToolTipLabel>
+          );
+        },
+      },
+      {
+        Header: "Quantity",
+        accessor: ({ quantity }: { quantity: string }) => {
+          return quantity || "1";
+        },
+      },
+    ] as Column<AirdropAddressInput>[];
+  }, []);
+
   return (
     <div className="flex w-full flex-col gap-6">
       {normalizeQuery.pending && <div>Loading... </div>}
-      {normalizeQuery.data?.result.length && validAirdrop.length > 0 ? (
+      {normalizeQuery.data?.result.length && rawData.length > 0 ? (
         <>
-          <AirdropTable
+          <CsvDataTable<AirdropAddressInput>
             portalRef={paginationPortalRef}
             data={normalizeQuery.data.result}
+            columns={columns}
           />
           <div className="mt-4 flex flex-col justify-between md:mt-0">
             <div ref={paginationPortalRef} />
             <div className="mt-3 ml-auto flex w-full flex-row gap-2 md:w-auto">
               <Button
                 borderRadius="md"
-                disabled={validAirdrop.length === 0}
+                disabled={rawData.length === 0}
                 onClick={() => {
                   reset();
                 }}
@@ -192,7 +113,7 @@ export const AirdropUpload: React.FC<AirdropUploadProps> = ({
                 <Button
                   borderRadius="md"
                   colorScheme="primary"
-                  disabled={validAirdrop.length === 0}
+                  disabled={rawData.length === 0}
                   onClick={() => {
                     removeInvalid();
                   }}
@@ -206,7 +127,7 @@ export const AirdropUpload: React.FC<AirdropUploadProps> = ({
                   colorScheme="primary"
                   onClick={onSave}
                   w={{ base: "100%", md: "auto" }}
-                  isDisabled={validAirdrop.length === 0}
+                  isDisabled={rawData.length === 0}
                 >
                   Next
                 </Button>
@@ -269,174 +190,5 @@ export const AirdropUpload: React.FC<AirdropUploadProps> = ({
         </div>
       )}
     </div>
-  );
-};
-
-interface AirdropTableProps {
-  data: AirdropAddressInput[];
-  portalRef: React.RefObject<HTMLDivElement | null>;
-}
-const AirdropTable: React.FC<AirdropTableProps> = ({ data, portalRef }) => {
-  const columns = useMemo(() => {
-    return [
-      {
-        Header: "Address",
-        accessor: ({ address, isValid }) => {
-          if (isValid) {
-            return address;
-          }
-          return (
-            <ToolTipLabel
-              label={
-                address === ZERO_ADDRESS
-                  ? "Cannot send tokens to ZERO_ADDRESS"
-                  : address.startsWith("0x")
-                    ? "Address is not valid"
-                    : "Address couldn't be resolved"
-              }
-            >
-              <div className="flex flex-row items-center gap-2">
-                <CircleAlertIcon className="size-4 text-red-500" />
-                <div className="cursor-default font-bold text-red-500">
-                  {address}
-                </div>
-              </div>
-            </ToolTipLabel>
-          );
-        },
-      },
-      {
-        Header: "Quantity",
-        accessor: ({ quantity }: { quantity: string }) => {
-          return quantity || "1";
-        },
-      },
-    ] as Column<AirdropAddressInput>[];
-  }, []);
-  const {
-    getTableProps,
-    getTableBodyProps,
-    headerGroups,
-    prepareRow,
-    // Instead of using 'rows', we'll use page,
-    page,
-    // which has only the rows for the active page
-    // The rest of these things are super handy, too ;)
-    canPreviousPage,
-    canNextPage,
-    pageOptions,
-    pageCount,
-    gotoPage,
-    nextPage,
-    previousPage,
-    setPageSize,
-    state: { pageIndex, pageSize },
-  } = useTable(
-    {
-      columns,
-      data,
-      initialState: {
-        pageSize: 50,
-        pageIndex: 0,
-      },
-    },
-    // old package: this will be removed
-    // eslint-disable-next-line react-compiler/react-compiler
-    usePagination,
-  );
-  // Render the UI for your table
-  return (
-    <>
-      <Table {...getTableProps()}>
-        <Thead>
-          {headerGroups.map((headerGroup, headerGroupIndex) => (
-            // biome-ignore lint/suspicious/noArrayIndexKey: FIXME
-            <Tr {...headerGroup.getHeaderGroupProps()} key={headerGroupIndex}>
-              {headerGroup.headers.map((column, columnIndex) => (
-                <Th
-                  {...column.getHeaderProps()}
-                  border="none"
-                  // biome-ignore lint/suspicious/noArrayIndexKey: FIXME
-                  key={columnIndex}
-                >
-                  <Text as="label" size="label.sm" color="faded">
-                    {column.render("Header")}
-                  </Text>
-                </Th>
-              ))}
-            </Tr>
-          ))}
-        </Thead>
-        <Tbody {...getTableBodyProps()}>
-          {page.map((row, rowIndex) => {
-            prepareRow(row);
-            return (
-              // biome-ignore lint/suspicious/noArrayIndexKey: FIXME
-              <Tr {...row.getRowProps()} key={rowIndex}>
-                {row.cells.map((cell, cellIndex) => (
-                  <Td
-                    {...cell.getCellProps()}
-                    borderColor="borderColor"
-                    // biome-ignore lint/suspicious/noArrayIndexKey: FIXME
-                    key={cellIndex}
-                  >
-                    {cell.render("Cell")}
-                  </Td>
-                ))}
-              </Tr>
-            );
-          })}
-        </Tbody>
-      </Table>
-      {/* Only need to show the Pagination components if we have more than 25 records */}
-      {data.length > 0 && (
-        <Portal containerRef={portalRef}>
-          <div className="flex w-full items-center justify-center">
-            <div className="flex flex-row gap-1">
-              <IconButton
-                isDisabled={!canPreviousPage}
-                aria-label="first page"
-                icon={<ChevronFirstIcon className="size-4" />}
-                onClick={() => gotoPage(0)}
-              />
-              <IconButton
-                isDisabled={!canPreviousPage}
-                aria-label="previous page"
-                icon={<ChevronLeftIcon className="size-4" />}
-                onClick={() => previousPage()}
-              />
-              <p className="my-auto whitespace-nowrap">
-                Page <strong>{pageIndex + 1}</strong> of{" "}
-                <strong>{pageOptions.length}</strong>
-              </p>
-              <IconButton
-                isDisabled={!canNextPage}
-                aria-label="next page"
-                icon={<ChevronRightIcon className="size-4" />}
-                onClick={() => nextPage()}
-              />
-              <IconButton
-                isDisabled={!canNextPage}
-                aria-label="last page"
-                icon={<ChevronLastIcon className="size-4" />}
-                onClick={() => gotoPage(pageCount - 1)}
-              />
-              <Select
-                onChange={(e) => {
-                  setPageSize(Number.parseInt(e.target.value as string, 10));
-                }}
-                value={pageSize}
-              >
-                <option value="25">25</option>
-                <option value="50">50</option>
-                <option value="100">100</option>
-                <option value="250">250</option>
-                <option value="500">500</option>
-              </Select>
-            </div>
-          </div>
-        </Portal>
-      )}
-    </>
   );
 };
